refactor: move Accordion back out of ArchivedComponents
