--- src/components/ArchivedComponents/Accordion/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./Accordion";
