+++ src/components/CurriculumComponents/CurriculumUnitDetailsAccordion/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./CurriculumUnitDetailsAccordion";
